markActiveTrack in case of default languaga is selected again
In case default Text Track defined in the manifest and the playback.textLanguage is the same, we need to make sure that that's the only one that is set as active in the TextTrack module.

The reason is - because it is the default in the manifest, its selected by the video element when it's loaded. After the player tries to select the same text track - no event is sent (because there is no change in the video text tracks).

Until now - this behaviour was OK. however, as we have the external captions default attribute, we just need to make sure only the final and selected text track is set as active.