feat: engine decorator (#324)
Engine decorator framework which will serve DAI use-cases.
The engine decorator will catch any API call to the engine and will decide what do to depends on its logic.
Plugin who wants to implement an engine decorator will need to have the signature:
```
getEngineDecorator(engine) { 
	...
}
```
The BaseEngineDecorator redirects each API call to the engine itself, so if plugin won't implement certain API call, by default it will bubble to the engine.
When an engine is chosen, the player will check if certain plugin implemented getEngineDecorator. If such exists, it will pass it the engine and will get the decorator instead.