fix(FEC-10844): advanced caption settings custom caption is always marked even when choose sample
Issue: condition of || doesn't work properly with 0 as current value and always take the second option.
Solution: create the defaultValues as static getter to be able to use it on static methods and on first initilize of the properties.
merge the setting and default value to make sure all set correctly.
add relevant tests.