fix(FEC-9370): safari video starts from beginning, not from startTime
Issue: there are videos like ima dai sample, video tag fires duration change twice and disable the seek action because of that.
Solution: for every duration change(should be 1 for media) we'll make the seek, for this media it'll make the seek after the second duration change and then we'll be in the correct place.