fix(FEC-11021): safari doesn't change the ended state on video tag when video is ended
Issue: today we're handling ended of playback in safari with timeout and then dispatching an error, then the state on video tag still not ended.
Solution: Before error trigger fake ended and fake the getter to response this state if it's exist, then all browser behave on the same way.