fix(FEC-13227): video starts unmuted in playlist after playing YT entry (#717)
### Description of the Changes

- bug fix

**the issue:**
when playing a playlist and the audio should be muted- after playing a YT entry, the player is becoming unmuted for the next entry.

**root cause:**
it seems that the playkit is checking the YT engine's `muted` value, while the youtube engine hasn't even loaded yet. Because of that, the autoplay handler is executing `onFallbackMutedAutoPlay` function instead of `onMutedAutoPlay` function. As a result, when playing a different entry after playing the YT one, the UI is unmuting the player- because of the `fallbackMutedAutoPlay` state.

**solution:**
at the point when we're checking the `muted` value, it is possible that the engine hasn't loaded yet (like in our case), so the muted value that we get is not reflecting the actual state of muted; the engine is returning false by default.
to handle this case, we need to check the muted value also in `_playbackAttributesState` and in the config itself, prior to checking the engine.

Solves FEC-13227