fix(FEC-13481): YouTube entries inside playlist cause wrong behavior (#818)
### Description of the Changes

**ths issue:**
in a playlist, when switching between "regular" videos and YT videos, at
some point when clicking on the video area of a yt video, for play/pause
actions, it doesn't react.

**root cause:**
there are duplicates of engine elements for youtube - which grabs the
clicks and prevent from play/pause to take any affect.
every time when switching from a non-youtube video to a youtube video,
it adds a yt-engine element under the playkit-container.

**solution:**
when switching engines, in case of a youtube video - cleanup yt engine
element before appending the current yt engine element.

Solves FEC-13481

### CheckLists

- [ ] changes have been done against master branch, and PR does not
conflict
- [ ] new unit / functional tests have been added (whenever applicable)
- [ ] test are passing in local environment
- [ ] Travis tests are passing (or test results are not worse than on
master branch :))
- [ ] Docs have been updated