fix(FEC-14684): Add fontweight option and adjust fontsize (#838)
This fixes this https://kaltura.atlassian.net/browse/FEC-14684 by
completing this https://github.com/kaltura/playkit-js-ui/pull/1095 with
fontWeight option and adjusting fonstSize based on product requests
### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

### CheckLists

- [ ] changes have been done against master branch, and PR does not
conflict
- [ ] new unit / functional tests have been added (whenever applicable)
- [ ] test are passing in local environment
- [ ] Travis tests are passing (or test results are not worse than on
master branch :))
- [ ] Docs have been updated