fix(FEC-14814): Switching between unidentify audio language to another doesn't work properly. (#845)
### Description of the Changes

Fix unknown language from playkit side instead of ui side.
This is a better and more localized fix than making multiple fixes in
the ui.
If the audio track's language code is "un", we try to calculate the
actual language code from the track label.
My only concern was that this wouldn't work because the manifest still
contains "un", but the track change does work.

Resolves FEC-14814, FEC-14815, FEC-14816 (and the original ticket
SUP-50865)